+++ addon/utils/plugins/br-skipping/backspace-plugin.ts
@@ -1,7 +1,7 @@
 import {
   BackspaceHandlerManipulation,
   BackspacePlugin
-} from '@lblod/ember-rdfa-editor/editor/input-handlers/backspace-handler';
+} from '@lblod/ember-rdfa-editor/editor/input-handlers/backspace-handler-old';
 import {
   ManipulationGuidance,
   MoveCursorToEndOfElementManipulation
