+++ addon/utils/plugins/empty-element/backspace-plugin.ts
@@ -1,7 +1,7 @@
 import {
   BackspaceHandlerManipulation,
   BackspacePlugin
-} from "@lblod/ember-rdfa-editor/editor/input-handlers/backspace-handler";
+} from "@lblod/ember-rdfa-editor/editor/input-handlers/backspace-handler-old";
 import { Editor,
          ManipulationGuidance,
          MoveCursorToEndOfElementManipulation} from "@lblod/ember-rdfa-editor/editor/input-handlers/manipulation";
